Hrc files are great, but for big clouds their creation makes the conversion very slow, because all of them are recreated at each flush. I have a cloud with about 3 million bin files and 300k hrc files, and their creation takes about half of the time.

With this patch:
* the hierarchy is not persisted at every flush, but more and more rarely, proceeding with the conversion
* the subtrees that are not modified since the last flush do not update their hierarchy.

It seems to work for me.