@m-schuetz 2015-08-12
Thanks for pointing out this issue. The PR is no longer compatible with the current branch but I'm going to use it as a reference for an up to date implementation.
I'd still like to flush the hierarchy at every flush, if possible. Hopefully it's fast enough if only modified hrc files are written.