@m-schuetz 2015-08-12
Pushed something now: https://github.com/potree/PotreeConverter/commit/8ecacc95db44348a340e8e0dfd20979d91cbaf20 (experimental)

hrc files are only written to disk if they contain at least one new node. 

After converting ~200m points, I get something like this:
´´´
hrcTotal: 16,659; hrcFlushed: 488
writing hierarchy: 0.018s
´´´

Previously 16.659 hrc files would have been written at that point and now it's just 488. 
Test is still running. Previously, flushing took 70s after 3 billion points.