@m-schuetz 2015-08-13
After converting 13 billion points with flush after every 50 million points:

```
INDEXING: 13,549,000,000 points processed; 13,549,000,000 points written; 41,322 seconds passed
INDEXING: 13,550,000,000 points processed; 13,550,000,000 points written; 41,323.7 seconds passed
FLUSHING: flush nodes: 42.498s
hrcTotal: 1,145,911; hrcFlushed: 4,345
writing hierarchy: 3.148s
46.683s
```

Hierarchy write duration for the last few flushes at >13b points ranged between ~1.5s to 6s. 
Writing nodes took 30s to 50s.

Hierarchy write duration at <1b ranged between 0.01s to 1s
Writing nodes took 20s to 40s.

So flush performance still decreases but not that bad anymore.